Enabling/Disabling hitboxes
Individual hitboxes can now be enabled for an attack and disabled
afterwards.

This means that each hitbox no longer needs a global variable. Disabled
hitboxes are also not calculated by the physics engine, improving
performance and meaning that enemies won't collide with hitboxes that
should be inactive.

Currently an issue where a hitbox will appear in the corner of the
screen for 1 frame when it is reset. I tested this and found that the
hitbox is active and will hit an enemy if one is present in the corner
during that frame.