Moved player to prototype
Created a prototype for player (in a new file). Moved player related
functions to this new file, cutting the size of the main game file by
more than half.

Minimal testing has been done on the moved code. Most things work, but
it is possible that bugs exist.

Still need to move the player hitboxes out of the main file. They are
currently still global.